Removed the hook for the sidebar links, moved the list of hooks to the sidebar template tag.
