--- tapir/core/config.py
@@ -1,1 +1,0 @@
-sidebar_links_providers = []
