--- tapir/supercoop_berlin/__init__.py
@@ -1,1 +1,0 @@
-from .views import *
