--- tapir/supercoop_berlin/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-
-from tapir.supercoop_berlin import views
-
-app_name = "supercoop_berlin"
-urlpatterns = []
